+++ PYTHON/calculator.py
@@ -9,7 +9,7 @@ import msvcrt
 # NEWIT core ref имопртируем только АЛУ (в котором уже есть регистры)
 # from registers import Registry
 import flags
-import ALU
+import cpu
 
 class Calculator:
 	""" ОСНОВНОЙ КЛАСС КАЛЬКУЛЯТОРА """
@@ -28,7 +28,7 @@ class Calculator:
 		self.flags = flags.Flags()  # флаги калькулятора
 		self.mode = mode      # включение/выключения отладочного режима
 		# в АЛУ добавляем ссылку на флаги, но сам АЛУ их не изменеят
-		self.__ALU = ALU.ALU(self.flags) # инициализация АЛУ
+		self.__CPU = cpu.ALU(self.flags) # инициализация АЛУ
 		print("press 'q' to exit, 0-9 to enter value, 'ESC'-to clear, 'ENTER' to evaluate")
 		self.displayRegisters()
 
@@ -38,7 +38,7 @@ class Calculator:
 		# self.A.clear()
 		# self.B.clear()
 		# очищаем АЛУ (фактически только регистр Z на данном этапе)
-		self.__ALU.clear()
+		self.__CPU.clear()
 		self.OP = None
 		self.flags.clear()
 
@@ -46,16 +46,16 @@ class Calculator:
 	# NEWIT core ref изменение вывода регистров А и В
 	def displayRegisters(self):
 		if self.mode == 0:
-			print(f"A='{self.__ALU.A}'  ({self.OP})  B='{self.__ALU.B}'  Z='{self.__ALU.Z}'  SignZ='{self.__ALU.Z.sign}'  SignA='{self.__ALU.A.sign}'  EQ={int(self.flags.EQ)}  CD={int(self.flags.CD)}  CONST={int(self.flags.CONST)}")
+			print(f"A='{self.__CPU.A}'  ({self.OP})  B='{self.__CPU.B}'  Z='{self.__CPU.Z}'  SignZ='{self.__CPU.Z.sign}'  SignA='{self.__CPU.A.sign}'  EQ={int(self.flags.EQ)}  CD={int(self.flags.CD)}  CONST={int(self.flags.CONST)}")
 		elif self.mode == 1:
 			print("\r" + " "*50, end='\r')
-			print(f"A='{self.__ALU.A}'  ({self.OP})  B='{self.__ALU.B}'  EQ={int(self.flags.EQ)}  CD={int(self.flags.CD)}  CONST={int(self.flags.CONST)}", end="\r")
+			print(f"A='{self.__CPU.A}'  ({self.OP})  B='{self.__CPU.B}'  EQ={int(self.flags.EQ)}  CD={int(self.flags.CD)}  CONST={int(self.flags.CONST)}", end="\r")
 		elif self.mode == 2:
 			print("\r" + " "*80, end="")
-			print("\r" + str(self.__ALU.A), end='')
+			print("\r" + str(self.__CPU.A), end='')
 		# изменился вид вывода (для возможности добавления новых значений)
 		elif self.mode == 3:
-			return (f"A='{self.__ALU.A}'  ({self.OP})  B='{self.__ALU.B}'"
+			return (f"A='{self.__CPU.A}'  ({self.OP})  B='{self.__CPU.B}'"
 				f"  EQ={int(self.flags.EQ)}  CD={int(self.flags.CD)}  CONST={int(self.flags.CONST)}")
 
 	# нажата цифровая клавиша. Ввод значения в регистр A
@@ -64,9 +64,9 @@ class Calculator:
 	def pressedDigitalKey(self, c: str):
 		self.flags.equal_not_pressed()
 		if c == '\x08':
-			self.__ALU.BS()
+			self.__CPU.BS()
 		else:	
-			self.__ALU.input(c)
+			self.__CPU.input(c)
 			# NEWIT флаг заполнения регистра устанавливается, если 0 не является первым
 			if self.flags.IS_NEW_INPUT and c != '0':
 				self.flags.enable_reg_filling()
@@ -77,9 +77,9 @@ class Calculator:
 		# NEWIT ввод отрицательного числа в калькуляторе организован как операция 0 - number
 		self.flags.equal_not_pressed()
 		# NEWIT опять вводим метод для обрезки незначащих 0 дробной части
-		self.__ALU.truncate()
+		self.__CPU.truncate()
 		# NEWIT core ref алгоритм выбора теперь полностью в АЛУ
-		self.__ALU.process(self.OP)
+		self.__CPU.process(self.OP)
 		# if self.flags.IS_OPERATON_POSSIBLE:
 		# 	self.__ALU.process(self.A, self.B, self.OP)
 		# self.B.copyFrom(self.A)
@@ -90,9 +90,9 @@ class Calculator:
 	# нажата клавиша "равно" - обработаем
 	def pressedEqual(self):
 		self.flags.equal_pressed()
-		self.__ALU.truncate()
+		self.__CPU.truncate()
 		# NEWIT core ref алгоритм выбора теперь полностью в АЛУ
-		self.__ALU.process(self.OP)
+		self.__CPU.process(self.OP)
 		# if self.flags.IS_OPS_CONTINUES:
 		# 	self.__ALU.process(self.A, self.B, self.OP)
 		# else:
