+++ PYTHON/cpu.py
